+++ rapidhash/src/fast.rs
@@ -24,64 +24,43 @@ use crate::inner;
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::quality::RapidHasher] for a higher quality hash output where necessary.
-pub type RapidHasher = inner::RapidHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
+pub type RapidHasher<'s> = inner::RapidHasher<'s, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
+
+/// A rapidhash equivalent to [std::hash::RandomState] that uses a random seed and secrets for
+/// minimal DoS resistance.
+///
+/// This initialises a [crate::quality::RapidHasher] with the following settings:
+/// - `AVALANCHE` is disabled.
+/// - `SPONGE` is enabled.
+/// - `COMPACT` is disabled.
+/// - `PROTECTED` is disabled.
+///
+/// Use [crate::quality::RandomState] for a higher quality but slower hash output where desirable.
+pub type RandomState = inner::RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A [std::hash::BuildHasher] trait compatible hasher that uses the [RapidHasher] algorithm.
 ///
-/// This is an alias for [inner::RapidBuildHasher] with the following settings:
+/// This initialises a [crate::quality::RapidHasher] with the following settings:
 /// - `AVALANCHE` is disabled.
 /// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
-/// Note there that [crate::RapidRandomState] should be used
-pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-/// A rapidhash equivalent to [std::hash::RandomState] that uses a random seed and secrets.
-pub type RandomState = inner::RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-/// A [std::collections::HashMap] that uses the [RapidHasher] hash.
-///
-/// This is an alias for [inner::RapidHashMap] with the following settings:
-/// - `AVALANCHE` is disabled.
-/// - `SPONGE` is enabled.
-/// - `COMPACT` is disabled.
-/// - `PROTECTED` is disabled.
-///
-/// Use [crate::quality::RapidHashMap] where higher hash collision resistance is required.
-#[cfg(any(feature = "std", docsrs))]
-pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-/// A [std::collections::HashSet] that uses the [RapidHasher] hash.
-///
-/// This is an alias for [inner::RapidHashSet] with the following settings:
-/// - `AVALANCHE` is disabled.
-/// - `SPONGE` is enabled.
-/// - `COMPACT` is disabled.
-/// - `PROTECTED` is disabled.
-///
-/// Use [crate::quality::RapidHashSet] where higher hash collision resistance is required.
-#[cfg(any(feature = "std", docsrs))]
-pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
+/// Use [crate::quality::SeedableState] for a higher quality but slower hash output where desirable.
+pub type SeedableState<'secrets> = inner::SeedableState<'secrets, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 #[cfg(any(feature = "std", docsrs))]
-pub use inner::HashMapExt;
+#[deprecated(since = "0.4.0", note = "Please use the top-level rapidhash::RapidHashMap instead")]
+pub use crate::RapidHashMap;
+
 #[cfg(any(feature = "std", docsrs))]
-pub use inner::HashSetExt;
+#[deprecated(since = "0.4.0", note = "Please use the top-level rapidhash::RapidHashSet instead")]
+pub use crate::RapidHashSet;
 
-#[cfg(all(test, feature = "std"))]
-mod tests {
-    use super::*;
+#[cfg(any(feature = "std", docsrs))]
+#[deprecated(since = "0.4.0", note = "Please use the top-level rapidhash::HashMapExt instead")]
+pub use crate::HashMapExt;
 
-    #[test]
-    fn test_hashmap_new() {
-        let mut map = RapidHashMap::new();
-        map.insert("key", "value");
-    }
-
-    #[test]
-    fn test_hashset_new() {
-        let mut set = RapidHashSet::new();
-        set.insert("value");
-    }
-}
+#[cfg(any(feature = "std", docsrs))]
+#[deprecated(since = "0.4.0", note = "Please use the top-level rapidhash::HashSetExt instead")]
+pub use crate::HashSetExt;