+++ rapidhash/src/lib.rs
@@ -1,14 +1,14 @@
 #![cfg_attr(docsrs, doc = include_str!("../README.md"))]
 #![cfg_attr(not(docsrs), doc = "# Rapidhash")]
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![cfg_attr(docsrs, feature(doc_cfg_hide))]
 #![cfg_attr(docsrs, doc(cfg_hide(docsrs)))]
 
-// note: we don't #![deny(unsafe_code)] as seeding.rs uses unsafe code to initialise the static
-// secrets, but we're confident about the safety of that code (and it cannot interact with
-// user-generated values).
+#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(feature = "nightly", feature(likely_unlikely))]
+#![cfg_attr(feature = "nightly", feature(hasher_prefixfree_extras))]
+
 #![deny(missing_docs)]
 #![deny(unused_must_use)]
 #![allow(clippy::manual_hash_one)]
@@ -22,5 +22,11 @@ pub mod v3;
 pub mod inner;
 pub mod fast;
 pub mod quality;
+#[cfg(any(feature = "std", docsrs))]
+mod collections;
 
 pub mod rng;
+
+#[doc(inline)]
+#[cfg(any(feature = "std", docsrs))]
+pub use collections::*;