+++ rapidhash/src/quality.rs
@@ -22,65 +22,27 @@ use crate::inner;
 /// - `PROTECTED` is disabled.
 ///
 /// Use [crate::fast::RapidHasher] for a lower quality but faster hash output where desirable.
-pub type RapidHasher = inner::RapidHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-/// A [std::hash::BuildHasher] inspired by [rapidhash::v3::rapidhash_v3] with a focus on hash quality.
-///
-/// This is an alias for [inner::RapidBuildHasher] with the following settings:
-/// - `AVALANCHE` is enabled.
-/// - `SPONGE` is enabled.
-/// - `COMPACT` is disabled.
-/// - `PROTECTED` is disabled.
-///
-/// Use [crate::fast::RapidHasher] for a lower quality but faster hash output where desirable.
-pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
+pub type RapidHasher<'s> = inner::RapidHasher<'s, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
 /// A rapidhash equivalent to [std::hash::RandomState] that uses a random seed and secrets for
 /// minimal DoS resistance.
+///
+/// This initialises a [RapidHasher] with the following settings:
+/// - `AVALANCHE` is enabled.
+/// - `SPONGE` is enabled.
+/// - `COMPACT` is disabled.
+/// - `PROTECTED` is disabled.
+///
+/// Use [crate::fast::RandomState] for a lower quality but faster hash output where desirable.
 pub type RandomState = inner::RandomState<AVALANCHE, SPONGE, COMPACT, PROTECTED>;
 
-/// A [std::collections::HashMap] that uses the [RapidHasher] hash.
+/// A [std::hash::BuildHasher] trait compatible hasher that uses the [crate::fast::RapidHasher] algorithm.
 ///
-/// This is an alias for [inner::RapidHashMap] with the following settings:
+/// This initialises a [RapidHasher] with the following settings:
 /// - `AVALANCHE` is enabled.
 /// - `SPONGE` is enabled.
 /// - `COMPACT` is disabled.
 /// - `PROTECTED` is disabled.
 ///
-/// Use [crate::fast::RapidHashMap] where lower hash collision resistance is acceptable.
-#[cfg(any(feature = "std", docsrs))]
-pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-/// A [std::collections::HashSet] that uses the [RapidHasher] hash.
-///
-/// This is an alias for [inner::RapidHashSet] with the following settings:
-/// - `AVALANCHE` is enabled.
-/// - `SPONGE` is enabled.
-/// - `COMPACT` is disabled.
-/// - `PROTECTED` is disabled.
-///
-/// Use [crate::fast::RapidHashSet] where lower hash collision resistance is acceptable.
-#[cfg(any(feature = "std", docsrs))]
-pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, SPONGE, COMPACT, PROTECTED>;
-
-#[cfg(any(feature = "std", docsrs))]
-pub use inner::HashMapExt;
-#[cfg(any(feature = "std", docsrs))]
-pub use inner::HashSetExt;
-
-#[cfg(all(test, feature = "std"))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_hashmap_new() {
-        let mut map = RapidHashMap::new();
-        map.insert("key", "value");
-    }
-
-    #[test]
-    fn test_hashset_new() {
-        let mut set = RapidHashSet::new();
-        set.insert("value");
-    }
-}
+/// Use [crate::fast::SeedableState] for a lower quality but faster hash output where desirable.
+pub type SeedableState<'secrets> = inner::SeedableState<'secrets, AVALANCHE, SPONGE, COMPACT, PROTECTED>;